requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.7
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette>=0.27.0
typing_extensions==4.4.0
uvicorn==0.22.0
watchfiles==0.18.1
websockets==10.4
pymongo[srv]
arrow
httpx[http2]
workalendar
loguru
jinja2
pandas
panel
altair
croniter
python-multipart
sortedcontainers
slack_sdk